add github workflows

+++ build-npm.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env sh
 wails3 generate bindings -ts -d npm-package/bindings &&
-    tsc -t es2022 -m es2022 --declaration --outDir npm-package/dist npm-package/index.ts &&
-    npm pack ./npm-package/dist
+    tsc -t es2022 -m es2022 --declaration --outDir npm-package/dist npm-package/index.ts
 
-rm -rf npm-package/dist/*.js npm-package/dist/*.d.ts
+if [ "$1" = "pack" ]; then
+    npm pack ./npm-package/dist
+    rm -rf npm-package/dist/*.js npm-package/dist/*.d.ts
+fi
